+++ src/bootloader/stage3/mysty/int.hpp
@@ -2,22 +2,22 @@
 
 using uint8_t = unsigned char;
 using uint16_t = unsigned short;
-using uint32_t = unsigned long;
+using uint32_t = unsigned int;
 using uint64_t = unsigned long long;
 
 using int8_t = char;
 using int16_t = short;
-using int32_t = long;
+using int32_t = int;
 using int64_t = long long;
 
 #ifdef __INTELLISENSE__
-using size_t = unsigned int;
+using size_t = unsigned long;
 #elif defined __clang__
-using size_t = unsigned int;
+using size_t = unsigned long;
 #elif defined __SIZE_TYPE__
 using size_t = __SIZE_TYPE__;
 #else
-using size_t = unsigned int;
+using size_t = unsigned long;
 #endif
 
 template <typename T>
